Metric Collection

+++ Mactivision Mini-Games/Assets/Scripts/Metrics/TimeAccuracyEvent.cs
@@ -1,10 +1,10 @@
 // LinearVariableEvent class: designed to be consumed by LinearVeriableMetric class.
 public class TimeAccuracyEvent : AbstractMetricEvent {
 
-    // Time in milliseconds away from desired time
-    public int accuracy { get; }
+    // Time away from desired time
+    public float accuracy { get; }
 
-    public TimeAccuracyEvent(System.DateTime eventTime, int accuracy) : base(eventTime) {
+    public TimeAccuracyEvent(System.DateTime eventTime, float accuracy) : base(eventTime) {
         this.accuracy = accuracy;
     }
 }
